\name{coefmclust}
\alias{coefmclust}
\title{ coefmclust }
\description{
  Returns coefficients of mclust type estimate. This function
  is only used internally.
}
\usage{
coefmclust(nmix)
}
\arguments{
  \item{nmix}{ Object of type \code{flexmix}. }
}
\value{
  \item{cov}{Description of 'comp1'}
  \item{center}{Description of 'comp2'}
  \item{prior}{Description...}
}
\author{P. Neytchev, P. Filzmoser, R. Patnaik, A. Eisl and R. Boubela,
<\email{P.Filzmoser@tuwien.ac.at}>
\url{http://www.statistik.tuwien.ac.at/public/filz/}}

\note{
	This function is only used internally.
}

\seealso{ \code{\link{coef}}, \code{\link[modeltools]{parameters}} }


\keyword{ cluster }
\keyword{ models }
\keyword{ multivariate }
