\name{estimate}
\alias{estimate}
\alias{estimate,TLE-method}
\title{ Getter for estimate object of TLE-ojbects. }
\description{
  This is the \code{estimate} function for TLE objects. It returns
  the estimate slot of TLE objects.
}
\usage{
	estimate(object)
}
\arguments{
  \item{object}{ Object of class TLE }
}
\value{
  \item{object}{Estimate object that is returned by the model driver's estimate function.}
}

\author{P. Neytchev, P. Filzmoser, R. Patnaik, A. Eisl and R. Boubela,
<\email{P.Filzmoser@tuwien.ac.at}>
\url{http://www.statistik.tuwien.ac.at/public/filz/}}
\seealso{ \code{\link{TLE}} }
\keyword{models}
