\name{TLE-methods}
\docType{methods}
\alias{TLE-methods}
\alias{TLE,numeric,numeric,ANY,numeric,numeric,ANY,numeric,ANY,function,function-method}
\title{Methods for Function TLE in Package `tlemix'}
\description{
 Methods for function \code{TLE} in Package `tlemix'
}
\section{Methods}{
\describe{
	\item{kStar = "numeric", kTrim = "numeric", data = "ANY", nit = "numeric", msglvl = "numeric", result = "ANY", cit = "numeric", test = "ANY", Density = "function", Estimate = "function"}{ TLE method }
}}
\keyword{methods}
