\name{tlemix-package}
\alias{tlemix-package}
\alias{tlemix}
\docType{package}
\title{
 TLE - Robust Estimation of Mixture Models using the Fast-TLE algorithm.
}
\description{
  TLE implements a general framework for robust fitting
   of finite mixture models. Parameter estimation is performed 
   using the EM algorithm.
}
\details{
	\tabular{ll}{
	Package: \tab tlemix\cr
	Type: \tab Package\cr
	Version: \tab 1.0\cr
	Date: \tab 2008-06-12\cr
	License: \tab GPL-2\cr
	LazyLoad: \tab yes\cr
	}
}
\author{
	P. Neytchev, P. Filzmoser, R. Patnaik, A. Eisl and R. Boubela
	Maintainer: P. Filzmoser, peter.filzmoser@tuwien.ac.at
}
\references{ N. Neykov, P. Filzmoser, R. Dimova, and P. Neytchev. Robust fitting of mixtures using the trimmed likelihood estimator. Computational Statistics and Data Analysis, Vol. 17(3), pp. 299-308, 2007. }
\keyword{ package }
\seealso{ \code{\link[flexmix]{flexmix}}, \code{\link{TLE}} }
